+++ starkbank/__init__.py
@@ -1,4 +1,4 @@
-version = "2.17.0"
+version = "2.18.0"
 
 user = None
 language = "en-US"